+++ interfaces.go
@@ -160,6 +160,9 @@ type Device interface {
 
 	// Parent, Owner, MAC Addresses if needed
 
+	// CreatePhysicalInterface will create a physical interface for this machine.
+	CreatePhysicalInterface(CreatePhysicalInterfaceArgs) (Interface, error)
+
 	// Delete will remove this Device.
 	Delete() error
 }
@@ -196,9 +199,6 @@ type Machine interface {
 
 	// Start the machine and install the operating system specified in the args.
 	Start(StartArgs) error
-
-	// CreatePhysicalInterface will create a physical interface for this machine.
-	CreatePhysicalInterface(CreatePhysicalInterfaceArgs) (Interface, error)
 }
 
 // Space is a name for a collection of Subnets.
@@ -244,9 +244,9 @@ type Interface interface {
 
 	MACAddress() string
 	EffectiveMTU() int
-	Params() string
 
-	// Need to work out types for children, discovered, parents
+	// Params is a JSON field, and defaults to an empty string, but is almost
+	// always a JSON object in practice. Gleefully ignoring it until we need it.
 }
 
 // Link represents a network link between an Interface and a Subnet.
