+++ link.go
@@ -83,7 +83,10 @@ func link_2_0(source map[string]interface{}) (*link, error) {
 		"mode":   schema.String(),
 		"subnet": schema.StringMap(schema.Any()),
 	}
-	checker := schema.FieldMap(fields, nil) // no defaults
+	defaults := schema.Defaults{
+		"subnet": schema.Omit,
+	}
+	checker := schema.FieldMap(fields, defaults)
 	coerced, err := checker.Coerce(source, nil)
 	if err != nil {
 		return nil, WrapWithDeserializationError(err, "link 2.0 schema check failed")
@@ -92,9 +95,12 @@ func link_2_0(source map[string]interface{}) (*link, error) {
 	// From here we know that the map returned from the schema coercion
 	// contains fields of the right type.
 
-	subnet, err := subnet_2_0(valid["subnet"].(map[string]interface{}))
-	if err != nil {
-		return nil, errors.Trace(err)
+	var subnet *subnet
+	if value, ok := valid["subnet"]; ok {
+		subnet, err = subnet_2_0(value.(map[string]interface{}))
+		if err != nil {
+			return nil, errors.Trace(err)
+		}
 	}
 
 	result := &link{
