+++ machine.go
@@ -5,9 +5,7 @@ package gomaasapi
 
 import (
 	"encoding/base64"
-	"fmt"
 	"net/http"
-	"strings"
 
 	"github.com/juju/errors"
 	"github.com/juju/schema"
@@ -184,82 +182,6 @@ func (m *machine) Start(args StartArgs) error {
 	}
 	m.updateFrom(machine)
 	return nil
-}
-
-// CreatePhysicalInterfaceArgs is an argument struct for passing parameters to
-// the Machine.CreatePhysicalInterface method.
-type CreatePhysicalInterfaceArgs struct {
-	// Name of the interface (required).
-	Name string
-	// MACAddress is the MAC address of the interface (required).
-	MACAddress string
-	// VLAN is the untagged VLAN the interface is connected to (required).
-	VLAN VLAN
-	// Tags to attach to the interface (optional).
-	Tags []string
-	// MTU - Maximum transmission unit. (optional)
-	MTU int
-	// AcceptRA - Accept router advertisements. (IPv6 only)
-	AcceptRA bool
-	// Autoconf - Perform stateless autoconfiguration. (IPv6 only)
-	Autoconf bool
-}
-
-// Validate checks the required fields are set for the arg structure.
-func (a *CreatePhysicalInterfaceArgs) Validate() error {
-	if a.Name == "" {
-		return errors.NotValidf("missing Name")
-	}
-	if a.MACAddress == "" {
-		return errors.NotValidf("missing MACAddress")
-	}
-	if a.VLAN == nil {
-		return errors.NotValidf("missing VLAN")
-	}
-	return nil
-}
-
-// CreatePhysicalInterface implements Machine.
-func (m *machine) CreatePhysicalInterface(args CreatePhysicalInterfaceArgs) (Interface, error) {
-	if err := args.Validate(); err != nil {
-		return nil, errors.Trace(err)
-	}
-	params := NewURLParams()
-	params.Values.Add("name", args.Name)
-	params.Values.Add("mac_address", args.MACAddress)
-	params.Values.Add("vlan", fmt.Sprint(args.VLAN.ID()))
-	params.MaybeAdd("tags", strings.Join(args.Tags, ","))
-	params.MaybeAddInt("mtu", args.MTU)
-	params.MaybeAddBool("accept_ra", args.AcceptRA)
-	params.MaybeAddBool("autoconf", args.Autoconf)
-	result, err := m.controller.post(m.interfacesURI(), "create_physical", params.Values)
-	if err != nil {
-		if svrErr, ok := errors.Cause(err).(ServerError); ok {
-			switch svrErr.StatusCode {
-			case http.StatusNotFound, http.StatusConflict:
-				return nil, errors.Wrap(err, NewBadRequestError(svrErr.BodyMessage))
-			case http.StatusForbidden:
-				return nil, errors.Wrap(err, NewPermissionError(svrErr.BodyMessage))
-			case http.StatusServiceUnavailable:
-				return nil, errors.Wrap(err, NewCannotCompleteError(svrErr.BodyMessage))
-			}
-		}
-		return nil, NewUnexpectedError(err)
-	}
-
-	iface, err := readInterface(m.controller.apiVersion, result)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-
-	m.interfaceSet = append(m.interfaceSet, iface)
-	return iface, nil
-}
-
-// interfacesURI used to add interfaces for this machine. The operations
-// are on the nodes endpoint, not machines.
-func (m *machine) interfacesURI() string {
-	return strings.Replace(m.resourceURI, "machines", "nodes", 1) + "interfaces/"
 }
 
 func readMachine(controllerVersion version.Number, source interface{}) (*machine, error) {
